Update GS2ML
Use command line args for game exe and data.win paths.
Allows this to work with games that aren't WYS, which was hardcoded.

Adds an error handler to prevent error messages from getting lost.
Previously, with this and gmml, if there was an error in the c# of a mod, it would typically close the console before showing any useful error info.
This was bad for developers making mods, and players reporting issues.
Now, if there is an error, it will prompt the user to enter "y" before launching the game.
Any other input closes it.
This allows the developer or user to see the error message and stack trace for the error, to report or debug it.